Follow-up PR to https://github.com/ggerganov/llama.cpp/pull/4538 . Like with e.g. LLaMA matrix matrix multiplications are now always done on the GPU for batch size >= 32. Performance:

| GPU      | Model     | Test | -ngl | t/s master | t/s PR | Speedup |
|----------|-----------|------|------|------------|--------|---------|
| RTX 3090 | 8x7b q2_K | pp   |    0 |       1.53 |  84.62 |   55.31 |
| RTX 3090 | 8x7b q2_K | pp   |   11 |       2.33 | 115.30 |   49.48 |
| RTX 3090 | 8x7b q2_K | pp   |   22 |       4.87 | 180.51 |   37.07 |
| RTX 3090 | 8x7b q2_K | pp   |   33 |     366.89 | 371.05 |    1.01 |