@Dampfinchen 2023-12-20
Wow, incredible results. Really glad that you decide to continue working on this project! You are simply a machine!