@Igoorx 2023-12-20
That's a god's send! I tried the PR, but it seems to crash with context `8192`, it works fine with context `4096` though.
```
CUDA error 1 at G:\llama.cpp\mixtral-fast\ggml-cuda.cu:8909: invalid argument
current device: 0
GGML_ASSERT: G:\llama.cpp\mixtral-fast\ggml-cuda.cu:8909: !"CUDA error"
```