@JohannesGaessler 2023-12-20
I cannot reproduce the issue.