@JohannesGaessler 2023-12-20
@Igoorx or are you the commit 4f420ea that I pushed at first? That commit had a bug which I then fixed with a force push.