@Igoorx 2023-12-20
@JohannesGaessler I was using the latest commit, this is the command line I was using:
```
server.exe -m "C:\temp\mixtral-8x7b-instruct-v0.1.Q5_K_M.gguf" -c 8192 -t 5 -ngl 10
```
And I'm using `Cuda 12.1`, if that matters.