@Igoorx 2023-12-20
@JohannesGaessler Okay actually nevermind, apparently after I rebooted my PC that crash went away lol