@JohannesGaessler 2023-12-20
I've changed the synchronization logic, please check whether it works now.