@Igoorx 2023-12-21
@JohannesGaessler That didn't solve it but I noticed something quite peculiar: the RAM usage with 512 context is 36GB~, and with 4096 context is 37GB~ but with 8192 context it skyrockets past 64GB! My total RAM is 64GB so I guess it's running out of memory.