@Igoorx 2023-12-21
I can't reproduce this issue using other models or with the master branch using Mixtral, but fair enough, maybe this was a bug with ggml itself that by accident got triggered by your PR, or somehow this is a bug with `Cuda Toolkit 12.1` (since the one compiled binary in the Releases page is using the version `12.2`, or my computer somehow is broken, honestly I have no clue 🤷