@Igoorx 2023-12-21
Ok, here is an update. I compiled the master branch myself but the issue still didn't happen. But after investigating further, looks like it doesn't happen with small prompts so maybe that's why you aren't being able to reproduce.
Try to use this as your prompt:
[prompt.txt](https://github.com/ggerganov/llama.cpp/files/13733860/prompt.txt)