@Dampfinchen 2023-12-21
> I got a 4.14 speedup on prompt evaluation for `mixtral-8x7b-instruct-v0.1.Q4_K_M.gguf` with 20 GPU layers (RTX 3090).
> 
> master:
> 
> llama_print_timings: load time = 15375.89 ms llama_print_timings: sample time = 872.01 ms / 512 runs ( 1.70 ms per token, 587.15 tokens per second) llama_print_timings: prompt eval time = 97628.28 ms / 3200 tokens ( 30.51 ms per token, 32.78 tokens per second) llama_print_timings: eval time = 44791.42 ms / 511 runs ( 87.65 ms per token, 11.41 tokens per second) llama_print_timings: total time = 144716.09 ms
> 
> PR:
> 
> llama_print_timings: load time = 4879.19 ms llama_print_timings: sample time = 93.91 ms / 512 runs ( 0.18 ms per token, 5452.14 tokens per second) llama_print_timings: prompt eval time = 23535.50 ms / 3200 tokens ( 7.35 ms per token, 135.96 tokens per second) llama_print_timings: eval time = 44120.32 ms / 511 runs ( 86.34 ms per token, 11.58 tokens per second) llama_print_timings: total time = 69207.50 ms
> 
> Amazing PR, thank you.

Out of curiosity, how do these results compare to 13B?