@Dampfinchen 2023-12-23
Hmm. LostRuins told me he still gets around 90 ms/t Prompt Processing on his RTX 2060 which is around 10 token/s, so very, very slow. That strikes me as really odd, considering the numbers im reading here. The 3090 is faster yes, but not that much faster. 

I wonder if there is more potential here to make the pp speed closer to 13B speeds (which would mean around 5 ms/t).